python/gcp/requirements.txt
python-storybook-core
pydantic
fastapi
uvicorn
inflection